@@ -5,6 +5,9 @@
   "description": "Search Tailwind CSS colors to use the HEX or RGB color value in projects that don't use Tailwind CSS classes.",
   "icon": "icon-tw-logo.png",
   "author": "magicjaqk",
+  "contributors": [
+    "ridemountainpig"
+  ],
   "categories": [
     "Developer Tools",
     "Design Tools",
@@ -21,16 +24,17 @@
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.33.2"
+    "@raycast/api": "^1.79.0",
+    "culori": "^4.0.1"
   },
   "devDependencies": {
-    "@types/node": "~16.10.0",
-    "@types/react": "^17.0.28",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "prettier": "^2.5.1",
+    "@raycast/eslint-config": "^1.0.11",
+    "@types/culori": "^2.1.1",
+    "@types/node": "^20.8.10",
+    "@types/react": "^18.3.3",
+    "eslint": "^8.57.1",
+    "prettier": "^3.5.3",
+    "react": "^18.2.0",
     "typescript": "^4.4.3"
   },
   "scripts": {
