@@ -7,40 +7,26 @@ import {
   Form,
   Icon,
   List,
-  LocalStorage,
   Toast,
   closeMainWindow,
-  environment,
   popToRoot,
   showToast,
-  showHUD,
   confirmAlert,
   open,
   getSelectedFinderItems,
   Keyboard,
 } from "@raycast/api";
 
-import { usePromise } from "@raycast/utils";
-import { useEffect, useRef, useState } from "react";
-
 import { runAppleScript } from "run-applescript";
-
-import { searchSpotlight } from "./search-spotlight";
-import { FolderSearchPlugin, SpotlightSearchResult } from "./types";
-
-import {
-  loadPlugins,
-  folderName,
-  enclosingFolderName,
-  showFolderInfoInFinder,
-  copyFolderToClipboard,
-  maybeMoveResultToTrash,
-  lastUsedSort,
-  fixDoubleConcat,
-} from "./utils";
-
+import path from "node:path";
 import fse from "fs-extra";
-import path = require("node:path");
+
+import { SpotlightSearchResult } from "./types";
+
+import { folderName, showFolderInfoInFinder, copyFolderToClipboard, maybeMoveResultToTrash } from "./utils";
+
+import { useFolderSearch } from "./hooks/useFolderSearch";
+import { FolderListSection } from "./components/FolderListSection";
 
 // allow string indexing on Icons
 interface IconDictionary {
@@ -50,412 +36,199 @@ interface IconDictionary {
 const IconDictionaried: IconDictionary = Icon as IconDictionary;
 
 export default function Command() {
-  const [searchText, setSearchText] = useState<string>("");
-  const [pinnedResults, setPinnedResults] = useState<SpotlightSearchResult[]>([]);
-  const [selectedItemId, setSelectedItemId] = useState<string>("");
-  const [searchScope, setSearchScope] = useState<string>("");
-  const [isShowingDetail, setIsShowingDetail] = useState<boolean>(true);
-  const [results, setResults] = useState<SpotlightSearchResult[]>([]);
+  const {
+    searchText,
+    setSearchText,
+    results,
+    isQuerying,
+    isShowingDetail,
+    setIsShowingDetail,
+    searchScope,
+    setSearchScope,
+    selectedItemId,
+    pinnedResults,
+    plugins,
+    resultIsPinned,
+    toggleResultPinnedStatus,
+    removeResultFromPinnedResults,
+    movePinUp,
+    movePinDown,
+    hasCheckedPlugins,
+    hasCheckedPreferences,
+  } = useFolderSearch();
 
-  // hack to fix annoying double text during fallback. Typing helloworld results in helloworldhelloworld
-  let fixedText = "";
-  useEffect(() => {
-    fixedText = fixDoubleConcat(searchText);
+  const sendFinderSelectionToFolder = async (destinationFolder: string) => {
+    const selectedItems = await getSelectedFinderItems();
 
-    if (fixedText !== searchText) {
-      setSearchText(fixedText); // Update the state of searchText
+    if (selectedItems.length === 0) {
+      await showToast(Toast.Style.Failure, "No Finder selection to send");
+      return;
     }
-  }, [searchText]);
 
-  const [plugins, setPlugins] = useState<FolderSearchPlugin[]>([]);
+    for (const item of selectedItems) {
+      const sourceFileName = path.basename(item.path);
+      const destinationFile = path.join(destinationFolder, sourceFileName);
 
-  const [isQuerying, setIsQuerying] = useState<boolean>(false);
-  const [canExecute, setCanExecute] = useState<boolean>(false);
-
-  const [hasCheckedPlugins, setHasCheckedPlugins] = useState<boolean>(false);
-  const [hasCheckedPreferences, setHasCheckedPreferences] = useState<boolean>(false);
-
-  const abortable = useRef<AbortController>();
-
-  // check plugins
-  usePromise(
-    async () => {
-      const plugins = await loadPlugins();
-      setPlugins(plugins);
-    },
-    [],
-    {
-      onData() {
-        setHasCheckedPlugins(true);
-      },
-      onError() {
-        showToast({
-          title: "An Error Occured",
-          message: "Could not read plugins",
-          style: Toast.Style.Failure,
-        });
-
-        setHasCheckedPlugins(true);
-      },
-    }
-  );
-
-  // check prefs
-  usePromise(
-    async () => {
-      const maybePreferences = await LocalStorage.getItem(`${environment.extensionName}-preferences`);
-
-      if (maybePreferences) {
-        try {
-          return JSON.parse(maybePreferences as string);
-        } catch (_) {
-          // noop
-        }
-      }
-    },
-    [],
-    {
-      onData(preferences) {
-        setPinnedResults(preferences?.pinned || []);
-        setSearchScope(preferences?.searchScope || "");
-        setIsShowingDetail(preferences?.isShowingDetail);
-        setHasCheckedPreferences(true);
-      },
-      onError() {
-        showToast({
-          title: "An Error Occured",
-          message: "Could not read preferences",
-          style: Toast.Style.Failure,
-        });
-
-        setHasCheckedPreferences(true);
-      },
-    }
-  );
-
-  // perform search
-  usePromise(
-    searchSpotlight,
-    [
-      searchText,
-      searchScope,
-      abortable,
-      (result: SpotlightSearchResult) => {
-        setResults((results) => [result, ...results].sort(lastUsedSort));
-      },
-    ],
-    {
-      onWillExecute: () => {
-        setIsQuerying(true);
-        setCanExecute(false);
-      },
-      onData: () => {
-        setIsQuerying(false);
-      },
-      onError: (e) => {
-        if (e.name !== "AbortError") {
-          showToast({
-            title: "An Error Occured",
-            message: "Something went wrong. Try again.",
-            style: Toast.Style.Failure,
+      try {
+        const exists = await fse.pathExists(destinationFile);
+        if (exists) {
+          const overwrite = await confirmAlert({
+            title: "Overwrite the existing file?",
+            message: sourceFileName + " already exists in " + destinationFolder,
           });
+
+          if (overwrite) {
+            if (item.path === destinationFile) {
+              await showToast(Toast.Style.Failure, "The source and destination file are the same");
+              continue;
+            }
+            fse.moveSync(item.path, destinationFile, { overwrite: true });
+            await showToast(Toast.Style.Success, "Moved file " + path.basename(item.path) + " to " + destinationFolder);
+          } else {
+            await showToast(Toast.Style.Failure, "Cancelling move");
+          }
+        } else {
+          fse.moveSync(item.path, destinationFile);
+          await showToast(Toast.Style.Success, "Moved file " + sourceFileName + " to " + destinationFolder);
         }
 
-        setIsQuerying(false);
-      },
-      execute: hasCheckedPlugins && hasCheckedPreferences && canExecute && !!searchText,
-      abortable,
-    }
-  );
-
-  // save preferences
-  useEffect(() => {
-    (async () => {
-      if (!(hasCheckedPlugins && hasCheckedPreferences)) {
-        return;
+        open(destinationFolder);
+      } catch (e) {
+        console.error("ERROR " + String(e));
+        await showToast(Toast.Style.Failure, "Error moving file " + String(e));
       }
-
-      await LocalStorage.setItem(
-        `${environment.extensionName}-preferences`,
-        JSON.stringify({
-          pinned: pinnedResults,
-          searchScope,
-          isShowingDetail,
-        })
-      );
-    })();
-  }, [pinnedResults, searchScope, isShowingDetail]);
-
-  useEffect(() => {
-    (async () => {
-      abortable.current?.abort();
-
-      setResults([]);
-      setIsQuerying(false);
-
-      // short-circuit for 'pinned'
-      if (searchScope === "pinned") {
-        setResults(
-          pinnedResults.filter((pin) =>
-            pin.kMDItemFSName.toLocaleLowerCase().includes(searchText.replace(/[[|\]]/gi, "").toLocaleLowerCase())
-          )
-        );
-        setCanExecute(false);
-      } else {
-        setCanExecute(true);
-      }
-    })();
-  }, [searchText, searchScope]);
-
-  const resultIsPinned = (result: SpotlightSearchResult) => {
-    return pinnedResults.map((pin) => pin.path).includes(result.path);
-  };
-
-  const removeResultFromPinnedResults = (result: SpotlightSearchResult) => {
-    setPinnedResults((pinnedResults) => pinnedResults.filter((pinnedResult) => pinnedResult.path !== result.path));
-  };
-
-  const toggleResultPinnedStatus = (result: SpotlightSearchResult, resultIndex: number) => {
-    if (!resultIsPinned(result)) {
-      setPinnedResults((pinnedResults) => [result, ...pinnedResults]);
-    } else {
-      // extracted out so we can re-use for maybeMoveResultToTrash calls
-      removeResultFromPinnedResults(result);
     }
 
-    setSelectedItemId(`result-${resultIndex.toString()}`);
+    closeMainWindow();
+    popToRoot({ clearSearchBar: true });
   };
 
-  const movePinUp = (result: SpotlightSearchResult, resultIndex: number) => {
-    const newIndex = resultIndex - 1;
-
-    [pinnedResults[resultIndex], pinnedResults[newIndex]] = [pinnedResults[newIndex], pinnedResults[resultIndex]];
-
-    setPinnedResults([...pinnedResults]);
-  };
-
-  const movePinDown = (result: SpotlightSearchResult, resultIndex: number) => {
-    const newIndex = resultIndex + 1;
-
-    [pinnedResults[resultIndex], pinnedResults[newIndex]] = [pinnedResults[newIndex], pinnedResults[resultIndex]];
-
-    setPinnedResults([...pinnedResults]);
-  };
-
-  // re-usable for results and 'pinned/favourites'
-  const ListSection = (props: { title: string; results: SpotlightSearchResult[] }) => {
+  // Render actions for the folder list items
+  const renderFolderActions = (result: SpotlightSearchResult, resultIndex: number) => {
     return (
-      <List.Section title={props.title}>
-        {props.results.map((result: SpotlightSearchResult, resultIndex: number) => (
-          <List.Item
-            id={`result-${resultIndex.toString()}`}
-            key={resultIndex}
-            title={folderName(result)}
-            subtitle={!isShowingDetail ? enclosingFolderName(result) : ""}
-            icon={{ fileIcon: result.path }}
-            accessories={resultIsPinned(result) ? [{ icon: { source: Icon.Star, tintColor: Color.Yellow } }] : []}
-            detail={
-              <List.Item.Detail
-                metadata={
-                  <List.Item.Detail.Metadata>
-                    <List.Item.Detail.Metadata.Label title="Metadata" />
-                    <List.Item.Detail.Metadata.Label title="Name" text={result.kMDItemFSName} />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label title="Where" text={result.path} />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label title="Type" text={result.kMDItemKind} />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label
-                      title="Created"
-                      text={result.kMDItemFSCreationDate?.toLocaleString()}
-                    />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label
-                      title="Modified"
-                      text={result.kMDItemContentModificationDate?.toLocaleString()}
-                    />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label
-                      title="Last used"
-                      text={result.kMDItemLastUsedDate?.toLocaleString() || "-"}
-                    />
-                    <List.Item.Detail.Metadata.Separator />
-                    <List.Item.Detail.Metadata.Label
-                      title="Use count"
-                      text={result.kMDItemUseCount?.toLocaleString() || "-"}
-                    />
-                    <List.Item.Detail.Metadata.Separator />
-                  </List.Item.Detail.Metadata>
-                }
+      <ActionPanel title={folderName(result)}>
+        <Action.Open
+          title="Open"
+          icon={Icon.Folder}
+          target={result.path}
+          onOpen={() => popToRoot({ clearSearchBar: true })}
+        />
+        <Action.ShowInFinder
+          title="Show in Finder"
+          path={result.path}
+          onShow={() => popToRoot({ clearSearchBar: true })}
+        />
+        <Action
+          title="Send Finder selection to Folder"
+          icon={Icon.Folder}
+          shortcut={{ modifiers: ["cmd", "shift"], key: "s" }}
+          onAction={() => sendFinderSelectionToFolder(result.path)}
+        />
+        <Action.OpenWith
+          title="Open With..."
+          shortcut={{ modifiers: ["cmd"], key: "o" }}
+          path={result.path}
+          onOpen={() => popToRoot({ clearSearchBar: true })}
+        />
+        <Action
+          title="Show Info in Finder"
+          icon={Icon.Finder}
+          shortcut={{ modifiers: ["cmd"], key: "i" }}
+          onAction={() => showFolderInfoInFinder(result)}
+        />
+        <Action
+          title="Toggle Details"
+          icon={Icon.Sidebar}
+          shortcut={{ modifiers: ["cmd", "shift"], key: "d" }}
+          onAction={() => setIsShowingDetail(!isShowingDetail)}
+        />
+        <Action
+          title={!resultIsPinned(result) ? "Pin" : "Unpin"}
+          icon={!resultIsPinned(result) ? Icon.Star : Icon.StarDisabled}
+          shortcut={{ modifiers: ["cmd", "shift"], key: "p" }}
+          onAction={() => toggleResultPinnedStatus(result, resultIndex)}
+        />
+        {resultIsPinned(result) && (
+          <>
+            {resultIndex > 0 && (
+              <Action
+                title="Move Pin Up"
+                icon={Icon.ArrowUpCircle}
+                shortcut={Keyboard.Shortcut.Common.MoveUp}
+                onAction={() => movePinUp(result, resultIndex)}
               />
-            }
-            actions={
-              <ActionPanel title={folderName(result)}>
-                <Action.Open
-                  title="Open"
-                  icon={Icon.Folder}
-                  target={result.path}
-                  onOpen={() => popToRoot({ clearSearchBar: true })}
-                />
-                <Action.ShowInFinder
-                  title="Show in Finder"
-                  path={result.path}
-                  onShow={() => popToRoot({ clearSearchBar: true })}
-                />
-                <Action
-                  title="Send Finder selection to Folder"
-                  icon={Icon.Folder}
-                  shortcut={{ modifiers: ["cmd", "shift"], key: "s" }}
-                  onAction={async () => {
-                    const selectedItems = await getSelectedFinderItems();
-
-                    if (selectedItems.length === 0) {
-                      await showHUD(`⚠️  No Finder selection to send.`);
-                    } else {
-                      for (const item of selectedItems) {
-                        // Source path = item
-                        // Source file name = path.basename(item)
-                        //
-                        // Destination folder = result.path
-                        // Destination file = result.path + '/' + path.basename(item)
-
-                        const sourceFileName = path.basename(item.path);
-                        const destinationFolder = result.path;
-                        const destinationFile = result.path + "/" + path.basename(item.path);
-
-                        try {
-                          const exists = await fse.pathExists(destinationFile);
-                          if (exists) {
-                            const overwrite = await confirmAlert({
-                              title: "Ooverwrite the existing file?",
-                              message: sourceFileName + " already exists in " + destinationFolder,
-                            });
-
-                            if (overwrite) {
-                              if (item.path == destinationFile) {
-                                await showHUD("The source and destination file are the same");
-                              }
-                              fse.moveSync(item.path, destinationFile, { overwrite: true });
-                              await showHUD("Moved file " + path.basename(item.path) + " to " + destinationFolder);
-                            } else {
-                              await showHUD("Cancelling move");
-                            }
-                          } else {
-                            fse.moveSync(item.path, destinationFile);
-                            await showHUD("Moved file " + sourceFileName + " to " + destinationFolder);
-                          }
-
-                          open(result.path);
-                        } catch (e) {
-                          console.error("ERROR " + String(e));
-                          await showToast(Toast.Style.Failure, "Error moving file " + String(e));
-                        }
-                      }
-                    }
-
-                    closeMainWindow();
-                    popToRoot({ clearSearchBar: true });
-                  }}
-                />
-                <Action.OpenWith
-                  title="Open With..."
-                  shortcut={{ modifiers: ["cmd"], key: "o" }}
-                  path={result.path}
-                  onOpen={() => popToRoot({ clearSearchBar: true })}
-                />
-                <Action
-                  title="Show Info in Finder"
-                  icon={Icon.Finder}
-                  shortcut={{ modifiers: ["cmd"], key: "i" }}
-                  onAction={() => showFolderInfoInFinder(result)}
-                />
-                <Action
-                  title="Toggle Details"
-                  icon={Icon.Sidebar}
-                  shortcut={{ modifiers: ["cmd", "shift"], key: "d" }}
-                  onAction={() => setIsShowingDetail(!isShowingDetail)}
-                />
-                <Action
-                  title={!resultIsPinned(result) ? "Pin" : "Unpin"}
-                  icon={!resultIsPinned(result) ? Icon.Star : Icon.StarDisabled}
-                  shortcut={{ modifiers: ["cmd", "shift"], key: "p" }}
-                  onAction={() => toggleResultPinnedStatus(result, resultIndex)}
-                />
-                {props.title === "Pinned" && (
-                  <>
-                    {resultIndex > 0 && (
-                      <Action
-                        title="Move Pin Up"
-                        icon={Icon.ArrowUpCircle}
-                        shortcut={Keyboard.Shortcut.Common.MoveUp}
-                        onAction={() => movePinUp(result, resultIndex)}
-                      />
-                    )}
-                    {resultIndex < props.results.length - 1 && (
-                      <Action
-                        title="Move Pin Down"
-                        icon={Icon.ArrowDownCircle}
-                        shortcut={Keyboard.Shortcut.Common.MoveDown}
-                        onAction={() => movePinDown(result, resultIndex)}
-                      />
-                    )}
-                  </>
-                )}
-                <ActionPanel.Section>
-                  <Action.CopyToClipboard
-                    title="Copy Folder"
-                    shortcut={{ modifiers: ["cmd"], key: "." }}
-                    content={``}
-                    onCopy={() => copyFolderToClipboard(result)}
-                  />
-                  <Action.CopyToClipboard
-                    title="Copy Name"
-                    shortcut={{ modifiers: ["cmd", "shift"], key: "." }}
-                    content={folderName(result)}
-                  />
-                  <Action.CopyToClipboard
-                    title="Copy Path"
-                    shortcut={{ modifiers: ["cmd", "shift"], key: "," }}
-                    content={result.path}
-                  />
-                </ActionPanel.Section>
-                <ActionPanel.Section>
-                  <Action.CreateQuicklink
-                    title="Create Quicklink"
-                    icon={Icon.Link}
-                    shortcut={{ modifiers: ["cmd", "shift"], key: "l" }}
-                    quicklink={{ link: result.path, name: folderName(result) }}
-                  />
-                </ActionPanel.Section>
-                <ActionPanel.Section>
-                  <Action
-                    title="Move to Trash"
-                    style={Action.Style.Destructive}
-                    icon={{ source: Icon.Trash, tintColor: Color.Red }}
-                    shortcut={{ modifiers: ["ctrl"], key: "x" }}
-                    onAction={() => maybeMoveResultToTrash(result, () => removeResultFromPinnedResults(result))}
-                  />
-                </ActionPanel.Section>
-                <ActionPanel.Section title="Plugins">
-                  {plugins.map((plugin: FolderSearchPlugin, pluginIndex) => (
-                    <Action
-                      key={pluginIndex}
-                      title={plugin.title}
-                      icon={IconDictionaried[plugin.icon]}
-                      shortcut={{ ...plugin.shortcut }}
-                      onAction={() => {
-                        popToRoot({ clearSearchBar: true });
-                        closeMainWindow({ clearRootSearch: true });
-                        runAppleScript(plugin.appleScript(result));
-                      }}
-                    />
-                  ))}
-                </ActionPanel.Section>
-              </ActionPanel>
-            }
+            )}
+            {resultIndex < pinnedResults.length - 1 && (
+              <Action
+                title="Move Pin Down"
+                icon={Icon.ArrowDownCircle}
+                shortcut={Keyboard.Shortcut.Common.MoveDown}
+                onAction={() => movePinDown(result, resultIndex)}
+              />
+            )}
+          </>
+        )}
+        <ActionPanel.Section>
+          <Action.CopyToClipboard
+            title="Copy Folder"
+            shortcut={{ modifiers: ["cmd"], key: "." }}
+            content={``}
+            onCopy={() => copyFolderToClipboard(result)}
           />
-        ))}
-      </List.Section>
+          <Action.CopyToClipboard
+            title="Copy Name"
+            shortcut={{ modifiers: ["cmd", "shift"], key: "." }}
+            content={folderName(result)}
+          />
+          <Action.CopyToClipboard
+            title="Copy Path"
+            shortcut={{ modifiers: ["cmd", "shift"], key: "," }}
+            content={result.path}
+          />
+        </ActionPanel.Section>
+        <ActionPanel.Section>
+          <Action.CreateQuicklink
+            title="Create Quicklink"
+            icon={Icon.Link}
+            shortcut={{ modifiers: ["cmd", "shift"], key: "l" }}
+            quicklink={{ link: result.path, name: folderName(result) }}
+          />
+        </ActionPanel.Section>
+        <ActionPanel.Section>
+          <Action
+            title="Move to Trash"
+            style={Action.Style.Destructive}
+            icon={{ source: Icon.Trash, tintColor: Color.Red }}
+            shortcut={{ modifiers: ["ctrl"], key: "x" }}
+            onAction={() => maybeMoveResultToTrash(result, () => removeResultFromPinnedResults(result))}
+          />
+        </ActionPanel.Section>
+        <ActionPanel.Section title="Plugins">
+          {plugins.map(
+            (
+              plugin: {
+                title: string;
+                icon: string;
+                shortcut: Keyboard.Shortcut;
+                appleScript: (result: SpotlightSearchResult) => string;
+              },
+              pluginIndex: number
+            ) => (
+              <Action
+                key={pluginIndex}
+                title={plugin.title}
+                icon={IconDictionaried[plugin.icon]}
+                shortcut={{ ...plugin.shortcut }}
+                onAction={() => {
+                  popToRoot({ clearSearchBar: true });
+                  closeMainWindow({ clearRootSearch: true });
+                  runAppleScript(plugin.appleScript(result));
+                }}
+              />
+            )
+          )}
+        </ActionPanel.Section>
+      </ActionPanel>
     );
   };
 
@@ -483,9 +256,21 @@ export default function Command() {
       }
     >
       {!searchText ? (
-        <ListSection title="Pinned" results={pinnedResults} />
+        <FolderListSection
+          title="Pinned"
+          results={pinnedResults}
+          isShowingDetail={isShowingDetail}
+          resultIsPinned={resultIsPinned}
+          renderActions={renderFolderActions}
+        />
       ) : (
-        <ListSection title="Results" results={results} />
+        <FolderListSection
+          title="Results"
+          results={results}
+          isShowingDetail={isShowingDetail}
+          resultIsPinned={resultIsPinned}
+          renderActions={renderFolderActions}
+        />
       )}
     </List>
   );
